Closes #18 - A big refactor of the Validator
The new Validator includes an internal DSL (Domain Specific Language) for configuring it for different validation scenarios, like validating a request, nested arrays, scalar values, or any given object. It's a basic first version, but highly extensible, so it's easy to create new rules. The first iteraction was very specific to the Routes Array, now it's extremely flexible.