Closes #19 - Refactor App Container: implements PSR-11 DI container
This is a big commit, as this refactor affected almost all parts of the application. It implements a PSR-11 DI container to dynamically inject dependencies into objects at runtime, allowing us to lazy-load them as needed. This also allowed to drop Guzzle as the API tester, so we can play with the API directly and make integration tests without HTTP calls. We also dropped Mockery in favor of native PHPUnit mocks.