Closes #15 - Implement JWT issuance and validation
A JWT token is issued on user login, and verified via a middleware. Tokens can be null, so calls without a token will pass, but calls with tokens will be validated first. Specialized routes will check for token presence at controller level, but they do not validate, they assume if it's present it's already validated.